pybrowser: refactor with own Exception class
ui.app is now $BROWSER or None. This way we recognize and handle
textbrowsers correctly when they are set by the environment.

The webbrowser module will give a bogus $BROWSER more trust and turn
it into a generic instance. Therefore we catch this with
PybrowserError with the webbrowser.Error message.